requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
av==15.1.0
boto3==1.40.40
botocore==1.40.40
certifi==2025.8.3
charset-normalizer==3.4.3
dm-tree==0.1.9
dotenv==0.9.9
filelock==3.19.1
fsspec==2025.9.0
gast==0.6.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-dali-cuda120==1.51.2
nvidia-nccl-cu12==2.27.3
nvidia-nvcomp-cu12==5.0.0.6
nvidia-nvimgcodec-cu12==0.6.0.32
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvjpeg-cu12==12.4.0.76
nvidia-nvjpeg2k-cu12==0.9.0.43
nvidia-nvtiff-cu12==0.5.1.75
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
s3transfer==0.14.0
safetensors==0.6.2
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
